% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra algorithm is used to find the shortest path find. It is the one of great algorithm in computer science to find the shortest path.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{First argument data frame contains nodes in two vectors v1 and v2 and also the distance between those nodes in w vector}

\item{init_node}{This second argument contains the initial node of the tree. The starting point to find the shortest path using Dijkstra algorithm.}
}
\value{
This return function returns the vector of the shortest paths of nodes sequentially.
}
\description{
Dijkstra's algorithm is an algorithm for finding the shortest paths between nodes in a graph, which may represent, for example, road networks. It was conceived by computer scientist Edsger W. Dijkstra in 1956 and published three years later.
}
\examples{
wiki_graph <-
data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
           v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
           w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
