% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm is used to find greatest common divisor}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{This argument is a numeric only value argument as a first value to find the euclidean}

\item{b}{This is a second argument with only numeric value to find the euclidean}
}
\value{
This return the euclidean final result
}
\description{
In mathematics, the Euclidean algorithm or Euclid's algorithm, is an efficient method for computing the greatest common divisor of two integers numbers, the largest number that divides them both without a remainder. Find Wikipedia docs \href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Here}.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
