% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Eighteen nodes are in vector name v1 with end nodes are in vector name v2 and eighteen distances in w
This graph contain 18 nodes edges and their distances.}
\format{
Data frame contains 18 rows and 3 variables:
\describe{
  \item{v1}{node}
  \item{v2}{end node}
  \item{w}{distance}
  ...
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra}
}
\usage{
wiki_graph
}
\description{
Eighteen nodes are in vector name v1 with end nodes are in vector name v2 and eighteen distances in w
This graph contain 18 nodes edges and their distances.
}
\keyword{datasets}
